+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# [next]
+- Improvements in `seeAndMoveToPlayer`. Adds param `obstacleCollision` sugested by issue [303](https://github.com/RafaelBarbosatec/bonfire/issues/303)
+
 # [2.9.1]
 - improvements in `MoveToPositionAlongThePath`. Resolve issue [299](https://github.com/RafaelBarbosatec/bonfire/issues/299)
 - improvements in `followComponent`. Now return `true` if can move. Resolve issue [301](https://github.com/RafaelBarbosatec/bonfire/issues/301)