+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # [next]
-- Improvements in `seeAndMoveToPlayer`. Adds param `obstacleCollision` sugested by issue [303](https://github.com/RafaelBarbosatec/bonfire/issues/303)
+- Improvements in `seeAndMoveToPlayer`. Adds param `notCanMove` sugested by issue [303](https://github.com/RafaelBarbosatec/bonfire/issues/303)
 - Improvements in `positionsItselfAndKeepDistance`.
 
 # [2.9.1]