+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# [next]
+# [2.9.2]
 - Improvements in `seeAndMoveToPlayer`. Adds param `notCanMove` sugested by issue [303](https://github.com/RafaelBarbosatec/bonfire/issues/303)
 - Improvements in `positionsItselfAndKeepDistance`.